+++ packages/catalog_core/lib/catalog_core.dart
@@ -19,6 +19,7 @@ export 'src/ics.dart';
 export 'src/match.dart';
 export 'src/transfer.dart';
 export 'src/pair_code.dart';
+export 'src/partial_date.dart';
 export 'src/registry.dart';
 export 'src/savepoints.dart';
 export 'src/sqlite_compat.dart';